fixes deploy bug, and more react setup
