+++ app/templates/_react/src/app/App.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+export default React.createClass({
+  render () {
+    return (
+      <div className='app__wrapper'>
+        {this.props.children}
+      </div>
+    );
+  }
+});
