+++ app/templates/_react/src/app/Home/Home.jsx
@@ -21,6 +21,7 @@ class Home extends React.Component {
   render () {
     return (
       <div className='home__wrapper'>
+        <h1>{this.props.title}</h1>
       </div>
     );
   }
