+++ app/templates/_react/src/app/Home/HomeStore.jsx
@@ -6,8 +6,14 @@ class HomeStore {
   constructor () {
     this.bindActions(HomeActions);
 
-    this.state = { };
-  } 
+    this.state = {
+      title: 'My Home Page'
+    };
+  }
+
+  onSetTitle (title) {
+    this.setState({title: title});
+  }
 }
 
 export default alt.createStore(HomeStore, 'HomeStore');
